refactor: extract curtailment API mappers
